#32 create package-info and refactor package structure
