+++ src/main/java/com/elbraulio/ezload/batch/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * This package contains all <code>batch</code> additions. i.e.
+ * {@link com.elbraulio.ezload.model.batch.IntBatch} or
+ * {@link com.elbraulio.ezload.model.batch.StringBatch}. All of them depends
+ * on its type.
+ *
+ * @author Braulio Lopez (brauliop.3@gmail.com)
+ */
+package com.elbraulio.ezload.model.batch;
