+++ src/main/java/com/elbraulio/ezload/constrain/NoConstrain.java
@@ -1,4 +1,4 @@
-package com.elbraulio.ezload.model;
+package com.elbraulio.ezload.constrain;
 
 /**
  * Represent no constrain needed.
