+++ src/main/java/com/elbraulio/ezload/exception/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * This package contains all exception that <cod>ezload</cod> can throw.
+ *
+ * @author Braulio Lopez (brauliop.3@gmail.com)
+ */
+package com.elbraulio.ezload.exception;
