+++ src/main/java/com/elbraulio/ezload/logger/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * This package contains defaults logger implementations.
+ *
+ * @author Braulio Lopez (brauliop.3@gmail.com)
+ */
+package com.elbraulio.ezload.logger;
