+++ src/main/java/com/elbraulio/ezload/model/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * This package has representations of a sql table as
+ * {@link com.elbraulio.ezload.model.Column}.
+ *
+ * @author Braulio Lopez (brauliop.3@gmail.com)
+ */
+package com.elbraulio.ezload.model;
