+++ src/main/java/com/elbraulio/ezload/parse/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * This package contains the File parser.
+ *
+ * @author Braulio Lopez (brauliop.3@gmail.com)
+ */
+package com.elbraulio.ezload.parse;
