+++ src/main/java/com/elbraulio/ezload/sql/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * This Package contains SQL manipulation like
+ * {@link com.elbraulio.ezload.sql.Insert} or
+ * {@link com.elbraulio.ezload.sql.BuildSql}.
+ *
+ * @author Braulio Lopez (brauliop.3@gmail.com)
+ */
+package com.elbraulio.ezload.sql;
