+++ src/main/java/com/elbraulio/ezload/transform/ToInt.java
@@ -1,4 +1,4 @@
-package com.elbraulio.ezload.model.transform;
+package com.elbraulio.ezload.transform;
 
 /**
  * Transform input to Integer.
