+++ src/main/java/com/elbraulio/ezload/transform/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * This package contains transformations for values.
+ *
+ * @author Braulio Lopez (brauliop.3@gmail.com)
+ */
+package com.elbraulio.ezload.model.transform;
