+++ src/test/java/com/elbraulio/ezload/model/GenericColumnTest.java
@@ -1,7 +1,8 @@
 package com.elbraulio.ezload.model;
 
-import com.elbraulio.ezload.model.batch.StringBatch;
-import com.elbraulio.ezload.model.transform.ToString;
+import com.elbraulio.ezload.batch.StringBatch;
+import com.elbraulio.ezload.constrain.NoConstrain;
+import com.elbraulio.ezload.transform.ToString;
 import org.hamcrest.CoreMatchers;
 import org.hamcrest.MatcherAssert;
 import org.junit.Test;
