+++ src/test/java/com/elbraulio/ezload/parse/DefaultParserTest.java
@@ -2,9 +2,9 @@ package com.elbraulio.ezload.parse;
 
 import com.elbraulio.ezload.model.Column;
 import com.elbraulio.ezload.model.GenericColumn;
-import com.elbraulio.ezload.model.NoConstrain;
-import com.elbraulio.ezload.model.batch.IntBatch;
-import com.elbraulio.ezload.model.transform.ToInt;
+import com.elbraulio.ezload.constrain.NoConstrain;
+import com.elbraulio.ezload.batch.IntBatch;
+import com.elbraulio.ezload.transform.ToInt;
 import org.hamcrest.CoreMatchers;
 import org.hamcrest.MatcherAssert;
 import org.junit.Test;
